collision-handle: Remove common and organize into type-specific files, fix linking
