+++ cfiles/edgelist.c
@@ -1,8 +1,23 @@
 #include <stdlib.h>
 #include <string.h>
 #ifndef BUILDSTDLIB
-#include "commonFunctions.h"
+#include "lockedobject.h"
 #endif
+
+struct EdgeList* createEdgeList() {
+    struct EdgeList* edge_list = malloc(sizeof(struct EdgeList));
+    edge_list->head = NULL;
+    edge_list->tail = NULL;
+    return edge_list;
+}
+
+struct EdgeListItem* createEdgeListItem(struct Edge* e) {
+    struct EdgeListItem* item = malloc(sizeof(struct EdgeListItem));
+    item->edge = e;
+    item->next = NULL;
+    item->prev = NULL;
+    return item;
+}
 
 struct Edge* removeFirst_EL(struct EdgeList* edge_list) {
     struct EdgeListItem *head;
@@ -45,8 +60,69 @@ void prependEdge(struct EdgeList* edge_list, struct Edge* e) {
     return; 
 }
 
-/*
-int main(){
-    return 0;
+bool empty_EL(struct EdgeList* edge_list) {
+    struct EdgeListItem *current;
+    current = edge_list->head;
+    return (current == NULL);
 }
-*/
+
+int length_EL(struct EdgeList* edge_list) {
+    int length = 0;
+    struct EdgeListItem *current;
+    current = edge_list->head;
+    while (current != NULL) {
+        length++;
+        current = current->next;
+    }
+    return length;
+}
+
+int removeEdge(struct EdgeList* edge_list, struct Edge* e) {
+    struct EdgeListItem* head = edge_list->head;
+    struct EdgeListItem* prev = NULL;
+    struct EdgeListItem* next = NULL;
+    if(head == NULL) {
+        // list is empty 
+        return -1;
+    } else {
+        prev = head->prev;
+        while (head) {
+            if (edgeEquals(e, head->edge)) {
+                if (prev) {
+                    next = head->next;
+                    prev->next = next;
+                    if (next) {
+                        next->prev = prev;
+                    }
+                } else {
+                    next = head->next;
+                    edge_list->head = next;
+                    if (next) {
+                        next->prev = NULL;
+                    }
+                }
+                return 0;
+            }
+            prev = head;
+            head = head->next;
+        }
+        return -1;
+    }
+}
+
+void appendEdge(struct EdgeList* edge_list, struct Edge* e) {
+    struct EdgeListItem* new_last = createEdgeListItem(e);
+    if (!empty_EL(edge_list)) {
+        // if list not empty;
+        new_last->prev = edge_list->tail;
+        struct EdgeListItem *old_last = malloc(sizeof(struct EdgeListItem));
+        old_last = edge_list->tail; 
+        old_last->next = new_last;
+        edge_list->tail = new_last;
+    } else {
+        // if list is empty;
+        edge_list->head = new_last;
+        edge_list->tail = new_last;
+    }
+    return;
+}
